临时提交代码

+++ src/main/java/com/wyyl1/hi/application/orgmng/OrgDto.java
@@ -7,15 +7,15 @@ import java.time.LocalDateTime;
 @Data
 public class OrgDto {
 
-    private Long id;
-    private Long tenantId;
-    private Long superiorId;
+    private Integer id;
+    private Integer tenantId;
+    private Integer superiorId;
     private String orgTypeCode;
-    private Long leaderId;
+    private Integer leaderId;
     private String name;
     private String status;
     private LocalDateTime createdAt;
-    private Long createdBy;
+    private Integer createdBy;
     private LocalDateTime lastUpdatedAt;
-    private Long lastUpdatedBy;
+    private Integer lastUpdatedBy;
 }
